+++ src/test/java/org/folio/rest/impl/PatronResourceImplTest.java
@@ -35,6 +35,7 @@ import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
 import java.util.Objects;
+import java.util.UUID;
 import java.util.stream.Stream;
 
 import static io.restassured.RestAssured.given;
@@ -393,6 +394,21 @@ public class PatronResourceImplTest extends BaseResourceServiceTest {
   }
 
   @Test
+  final void testGetPatronRegistrationStatusByESID() {
+    given()
+      .header(tenantHeader)
+      .header(urlHeader)
+      .header(contentTypeHeader)
+      .when()
+      .get(patronAccountRegistrationStatus + "/cacc29d8-cade-4312-a5f2-4eeac55d8700")
+      .then()
+      .log().all()
+      .contentType(ContentType.JSON)
+      .statusCode(200)
+      .extract().response();
+  }
+
+  @Test
   final void testGetPatronRegistrationStatusByEmailMultipleError() {
     final Errors errors = given()
       .header(tenantHeader)
@@ -411,6 +427,24 @@ public class PatronResourceImplTest extends BaseResourceServiceTest {
   }
 
   @Test
+  final void testGetPatronRegistrationStatusByESIDMultipleError() {
+    final Errors errors = given()
+      .header(tenantHeader)
+      .header(urlHeader)
+      .header(contentTypeHeader)
+      .when()
+      .get(patronAccountRegistrationStatus + "/cacc29d8-cade-4312-a5f2-4eeac55d7900")
+      .then()
+      .log().all()
+      .contentType(JSON)
+      .statusCode(400)
+      .extract().response().as(Errors.class);
+
+    assertEquals(1, errors.getErrors().size());
+    assertEquals(MULTIPLE_USER_WITH_EMAIL.value(), errors.getErrors().get(0).getMessage());
+  }
+
+  @Test
   final void testGetPatronRegistrationStatusByEmailUserNotFoundError() {
     var res = given()
       .header(tenantHeader)
@@ -418,6 +452,24 @@ public class PatronResourceImplTest extends BaseResourceServiceTest {
       .header(contentTypeHeader)
       .when()
       .get(patronAccountRegistrationStatus + "/notfound@folio.com")
+      .then()
+      .log().all()
+      .contentType(JSON)
+      .statusCode(404)
+      .extract().response().as(Errors.class);
+
+    assertEquals(1, res.getErrors().size());
+    assertEquals(USER_NOT_FOUND.value(), res.getErrors().get(0).getMessage());
+  }
+
+  @Test
+  final void testGetPatronRegistrationStatusByESIDUserNotFoundError() {
+    var res = given()
+      .header(tenantHeader)
+      .header(urlHeader)
+      .header(contentTypeHeader)
+      .when()
+      .get(patronAccountRegistrationStatus + "/cacc29d8-cade-4312-a5f2-4eeac55d8900")
       .then()
       .log().all()
       .contentType(JSON)
@@ -1458,7 +1510,7 @@ public class PatronResourceImplTest extends BaseResourceServiceTest {
       .header(contentTypeHeader)
       .body(jsonObject.encode())
       .when()
-      .post("/patron")
+      .put("/patron/" + UUID.randomUUID().toString())
       .then()
       .contentType(JSON)
       .statusCode(200);
@@ -1498,6 +1550,24 @@ public class PatronResourceImplTest extends BaseResourceServiceTest {
       .statusCode(400)
       .body(containsString("A bad exception occurred"))
       .contentType(TEXT);
+  }
+
+  @Test
+  final void testFailure404PutStagingUser() {
+    String body = readMockFile(MOCK_DATA_FOLDER + "/staging-users-post-request.json");
+    JsonObject jsonObject = new JsonObject(body);
+    jsonObject.getJsonObject("generalInfo").put("firstName", "TEST_STATUS_CODE_404");
+    given()
+      .log().all()
+      .header(tenantHeader)
+      .header(urlHeader)
+      .header(contentTypeHeader)
+      .body(jsonObject.encode())
+      .when()
+      .put("/patron/" + UUID.randomUUID().toString())
+      .then()
+      .statusCode(404)
+      .extract().response().as(Errors.class);
   }
 
   @Test
@@ -1557,6 +1627,23 @@ public class PatronResourceImplTest extends BaseResourceServiceTest {
       .then()
       .statusCode(500)
       .contentType(TEXT);
+  }
+
+  @Test
+  final void testExceptionPutStagingUser() {
+    String body = readMockFile(MOCK_DATA_FOLDER + "/staging-users-post-request.json");
+    JsonObject jsonObject = new JsonObject(body);
+    jsonObject.getJsonObject("generalInfo").put("firstName", "TEST_EXCEPTIONALLY_PART");
+    given()
+      .log().all()
+      .header(tenantHeader)
+      .header(urlHeader)
+      .header(contentTypeHeader)
+      .body(jsonObject.encode())
+      .when()
+      .put("/patron")
+      .then()
+      .statusCode(405);
   }
 
   static Stream<Arguments> instanceHoldsFailureCodes() {
@@ -1831,16 +1918,32 @@ public class PatronResourceImplTest extends BaseResourceServiceTest {
             .setStatusCode(200)
             .putHeader("content-type", "application/json")
             .end(readMockFile(MOCK_DATA_FOLDER + "/user_active1.json"));
-        } else if (req.uri().equals("/users?query=%28personal.email%3D%3Dinactive%40folio.com%20and%20type%3D%3Dpatron%29")) {
+        } else if (req.uri().equals("/users?query=%28externalSystemId%3D%3Dcacc29d8-cade-4312-a5f2-4eeac55d8700%20and%20type%3D%3Dpatron%29")) {
+          req.response()
+            .setStatusCode(200)
+            .putHeader("content-type", "application/json")
+            .end(readMockFile(MOCK_DATA_FOLDER + "/user_active1.json"));
+        }
+        else if (req.uri().equals("/users?query=%28personal.email%3D%3Dinactive%40folio.com%20and%20type%3D%3Dpatron%29")) {
           req.response()
             .setStatusCode(200)
             .putHeader("content-type", "application/json")
             .end(readMockFile(MOCK_DATA_FOLDER + "/user_not_active1.json"));
+        } else if (req.uri().equals("/users?query=%28externalSystemId%3D%3Dcacc29d8-cade-4312-a5f2-4eeac55d8900%20and%20type%3D%3Dpatron%29")) {
+          req.response()
+            .setStatusCode(200)
+            .putHeader("content-type", "application/json")
+            .end(readMockFile(MOCK_DATA_FOLDER + "/user_not_found.json"));
         } else if (req.uri().equals("/users?query=%28personal.email%3D%3Dnotfound%40folio.com%20and%20type%3D%3Dpatron%29")) {
           req.response()
             .setStatusCode(200)
             .putHeader("content-type", "application/json")
             .end(readMockFile(MOCK_DATA_FOLDER + "/user_not_found.json"));
+        } else if (req.uri().equals("/users?query=%28externalSystemId%3D%3Dcacc29d8-cade-4312-a5f2-4eeac55d7900%20and%20type%3D%3Dpatron%29")) {
+          req.response()
+            .setStatusCode(200)
+            .putHeader("content-type", "application/json")
+            .end(readMockFile(MOCK_DATA_FOLDER + "/user_multiple.json"));
         } else if (req.uri().equals("/users?query=%28personal.email%3D%3Dmultiple%40folio.com%20and%20type%3D%3Dpatron%29")) {
           req.response()
             .setStatusCode(200)
@@ -2399,7 +2502,7 @@ public class PatronResourceImplTest extends BaseResourceServiceTest {
               .end("Invalid combination of query parameters");
           }
         }
-      } else if (req.method() == HttpMethod.POST && req.uri().equals("/staging-users")) {
+      } else if (req.method() == HttpMethod.PUT && req.uri().contains("/staging-users")) {
         req.bodyHandler(body -> {
           String content = new String(body.getBytes());
           JsonObject jsonContent = new JsonObject(content);
@@ -2409,7 +2512,35 @@ public class PatronResourceImplTest extends BaseResourceServiceTest {
               .setStatusCode(200)
               .putHeader("content-type", "application/json")
               .end(readMockFile(MOCK_DATA_FOLDER + "/staging-users-post-response.json"));
-          } else if ("TEST_STATUS_CODE_201".equals(firstName)) {
+          } else if ("TEST_STATUS_CODE_400".equals(firstName)) {
+            req.response()
+              .setStatusCode(400)
+              .putHeader("content-type", "text/plain")
+              .end("A bad exception occurred");
+          }
+          else if ("TEST_STATUS_CODE_404".equals(firstName)) {
+            req.response()
+              .setStatusCode(404)
+              .putHeader("content-type", "text/plain")
+              .end("Staging user not found");
+          } else if ("TEST_STATUS_CODE_422".equals(firstName)) {
+            req.response()
+              .setStatusCode(422)
+              .putHeader("content-type", "application/json")
+              .end(readMockFile(MOCK_DATA_FOLDER + "/staging-users-post-error-response.json"));
+          } else if ("TEST_STATUS_CODE_500".equals(firstName)) {
+            req.response()
+              .setStatusCode(500)
+              .putHeader("content-type", "text/plain")
+              .end("A random exception occurred");
+          }
+        });
+      } else if (req.method() == HttpMethod.POST && req.uri().equals("/staging-users")) {
+        req.bodyHandler(body -> {
+          String content = new String(body.getBytes());
+          JsonObject jsonContent = new JsonObject(content);
+          String firstName = jsonContent.getJsonObject("generalInfo").getString("firstName");
+          if ("TEST_STATUS_CODE_201".equals(firstName)) {
             req.response()
               .setStatusCode(201)
               .putHeader("content-type", "application/json")